backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.5
Pillow==10.2.0
drf-spectacular==0.27.2
